Don't check in the changelog, just hand it to GitHub (#19668)
This change primarily moves changelog generation from release prep
(`start_release.py`) to a dedicated script (`changelog.py`) which just
returns the relevant changelog entries given a tag.

The expected usage is exactly the change made to the `release` job,
which is to scrape and categorize the PR titles solely for the GitHub
release notes.

Changes to `start_release.py`:
- Split all changelog-related code into `changelog.py`
- Changed it to also work on release branches
- E.g. `pants run start_release -- 2.17.0rc5` will make a PR against
`2.17.x` with the version/contributor bump
- No longer comments on the PR. No need.

Changes to the changelog code split from `start_release.py`:
- Since we're running this post-tag, we can now ask for the tag prior to
this one
- I switched to `pygithub` for label scraping
- The headers were each decremented one level (from `###` to `##`)

`generate_github_workflows.py` now uses `changelog.py` and therefore
`get_release_notes.py` is gone.
